Move Meal planner into it's own file. You know, for orgainzation
