+++ lib/shoppingList.dart
@@ -1,0 +1,23 @@
+import 'dart:async';
+import 'package:flutter/material.dart';
+import 'package:flutter/src/widgets/framework.dart';
+
+//Shopping list page
+class ShoppingList extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text("Second Route"),
+      ),
+      body: Center(
+        child: ElevatedButton(
+          onPressed: () {
+            Navigator.pop(context);
+          },
+          child: Text('Go back!'),
+        ),
+      ),
+    );
+  }
+}
